requirements.txt
streamlit
langchain
langchain-groq
groq
python-dotenv
psutil